update: setTokens parameters
