+++ src/pages/signin.tsx
@@ -1,4 +1,4 @@
-import { useState } from "react";
+import { useEffect, useState } from "react";
 import { Form } from "../styles/layout/sign/globalSignBox";
 import Link from "next/link";
 import { useForm } from "react-hook-form";
@@ -6,10 +6,15 @@ import { SignType } from "../components/signup/SignUp";
 import VisibilityIcon from "@mui/icons-material/Visibility";
 import VisibilityOffIcon from "@mui/icons-material/VisibilityOff";
 import { useRouter } from "next/router";
-import { useTokenContext } from "../context/tokenState";
+import { useTokenContext } from "../context/TokenState";
 import useFetch from "../hooks/useFetch";
 
 interface LoginDto {
+  student_id: string;
+  password: string;
+}
+
+interface LoginReturnDto {
   access_token: string;
   refresh_token: string;
 }
@@ -18,24 +23,31 @@ function SignIn() {
   const router = useRouter();
   const [show, setShow] = useState(false);
   const { register, handleSubmit } = useForm<SignType>();
-  const { setToken } = useTokenContext();
+  const { refreshToken, setTokens } = useTokenContext();
   const { post } = useFetch(
     "http://ec2-3-35-104-193.ap-northeast-2.compute.amazonaws.com:8000"
   );
 
-  const onSubmit = async (data: SignType) => {
-    const { access_token, refresh_token, error } = await post<LoginDto>(
+  const onSubmit = async (data: LoginDto) => {
+    const { access_token, refresh_token, error } = await post<LoginReturnDto>(
       "/auth/login",
       data
     );
     if (error) {
       alert("로그인실패");
     } else {
-      setToken(access_token, refresh_token);
+      setTokens({ accessToken: access_token, refreshToken: refresh_token });
       alert("로그인성공");
       router.push("/");
     }
   };
+
+  useEffect(() => {
+    if (refreshToken) {
+      router.push("/");
+    }
+  }, []);
+
   return (
     <Form onSubmit={handleSubmit(onSubmit)} show={show}>
       <input
